alltagshelfer-be-main/backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
coverage==7.2.2
Django==4.1.7
django-cors-headers==3.14.0
django-cron==0.6.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
graphviz==0.20.1
gunicorn==20.1.0
mccabe==0.7.0
nose==1.3.7
psycopg2==2.9.5
pycodestyle==2.10.0
pydotplus==2.0.2
pyflakes==3.0.1
PyJWT==2.6.0
pyparsing==3.1.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
Unidecode==1.3.6
whitenoise==6.4.0
